+++ Dimension/Dimension_Impl/Conversion.h
@@ -5,28 +5,10 @@
 
 namespace Dimension
 {
-   /// Concept to check if T has a numeric slope attribute.
-   template<typename T>
-   concept HasSlope = requires { T::slope; } && std::is_arithmetic_v<decltype(T::slope)>;
 
-   /// Concept to check if T has a numeric offset attribute.
+   /// @brief Concept to check if T has a numeric offset attribute.
    template<typename T>
    concept HasOffset = requires { T::offset; } && std::is_arithmetic_v<decltype(T::offset)>;
-
-   /// @brief Return the slope as a constexpr if one exists,
-   ///    otherwise return 1.0
-   template<typename T>
-   constexpr PrecisionType GetSlope()
-   {
-      if constexpr (HasSlope<T>)
-      {
-         return T::slope;
-      }
-      else
-      {
-         return 1.0;
-      }
-   }
 
    /// @brief Return the offset as a constexpr if one exists,
    ///    otherwise return 0.0
@@ -44,101 +26,101 @@ namespace Dimension
       }
    }
 
-   /// @brief Conversion implementation
-   /// @details base case when no conversion is defined, throws a compile-time error.
-   template<typename fromUnit, typename toUnit, typename Enable = void>
-   struct ConversionBase
-   {
-      static_assert(sizeof(fromUnit) == -1, "Conversion not defined for these units.");
+   /// @brief Conversion traits to be defined for each conversion
+   /// @details For each conversion, define a slope and optionally an offset
+   /// @tparam From Unit to convert from
+   /// @tparam To Unit to convert to
+   template<typename From, typename To>
+   struct Conversion;
+
+   /// @brief Concept to check if Conversion is defined
+   template<typename From, typename To>
+   concept HasConversion = requires { Conversion<From, To>::slope; };
+
+   /// @brief Concept to check if the unit is convertible to a primary unit (either direction)
+   template<typename T>
+   concept PrimaryConvertible = requires {
+      typename T::Primary;
+      requires (std::is_same_v<T, typename T::Primary> || HasConversion<T, typename T::Primary> || HasConversion<typename T::Primary, T>);
    };
 
-   /// @brief Struct defining the linear relationship between two units
-   /// @details This relationship is meant to be specialized for each unit,
-   ///    and by users when extensions are needed.
-   template<typename fromUnit, typename toUnit>
-   struct Conversion : ConversionBase<fromUnit, toUnit> {};
-
-   /// @brief Struct defining the linear relationship between two units
-   /// @details This specialization will be used when a final specialization is not provided, but the dimensions
-   ///    are the same. This will step through the primary unit of this dimension.
-   /// @typedef slope The linear relationship between the units
-   /// @typedef offset The intercept between two units.
-   ///    This will typically be 0.0, but there are some special cases, such as temparature.
-   /// @todo Provide some insight to the user when this is used, so they may accurately
-   ///    define useful specializations.
-   template<typename fromUnit, typename toUnit>
-   requires std::is_same_v<typename fromUnit::Dim, typename toUnit::Dim>
-   struct ConversionBase<fromUnit, toUnit>
+   /// @brief Implementation of convert function
+   /// @tparam Conv Conversion struct
+   /// @tparam IsDelta Bool indicating if the unit being converted is sole unit of the dimension 
+   ///   and is in the numerator (False), otherwise True
+   /// @tparam Inverse Whether to use the inverse of the conversion traits
+   /// @param input value to convert
+   /// @return converted value
+   template<typename Conv, bool IsDelta, bool Inverse>
+   constexpr PrecisionType ConvertImpl(PrecisionType input)
    {
+      // Assumes all checks have been made
 
-      #ifdef REQUIRE_CONVERSIONS
-         static_assert(sizeof(fromUnit) == -1, "No specialized conversion found. See compiler output for more details");
-      #endif
-
-      using toPrimary = Conversion<fromUnit, typename fromUnit::Primary>;
-      using fromPrimary = Conversion<typename fromUnit::Primary, toUnit>;
-
-      static constexpr PrecisionType slope = GetSlope<toPrimary>() * GetSlope<fromPrimary>();
-      static constexpr PrecisionType offset = GetOffset<toPrimary>() + GetOffset<fromPrimary>();
-   };
-
-   /// @brief Struct defining the linear relationship between two units
-   /// @details This relationship is meant to be specialized for each unit,
-   ///    and by users when extensions are needed.
-   ///    This specialization is used when converting to the same unit.
-   template<typename Unit>
-   struct Conversion<Unit, Unit>
-   {
-      static constexpr PrecisionType slope = 1.0;
-      static constexpr PrecisionType offset = 0.0;
-   };
-
-   /// @brief Method to convert a value of fromUnit to a value of toUnit
-   /// @details This method relies on the slope and offset of the necessary conversion.
-   ///    While this method can be specialized to bypass the slope-offset relationship,
-   ///    this should be done with great care as other parts of this library make assumptions
-   ///    about this relationship.
-   /// @tparam fromUnit Unit to convert from
-   /// @tparam toUnit Unit to convert to
-   /// @tparam isDelata Bool indicating whether this conversion is of a single unit in the normator (false)
-   ///    or not (true).
-   /// @tparam inverse Indicates whether value is in denominator (true) or numerator (false)
-   /// @param[in] input Value to convert as a floating-point type
-   /// @return floating-point type after conversion
-   template<typename fromUnit, typename toUnit, bool isDelta = false, bool inverse = false>
-   PrecisionType Convert(PrecisionType input)
-   {
-      if constexpr (std::is_same_v<fromUnit, toUnit>)
+      if constexpr (!Inverse)
       {
-         return input;
-      }
-      else if constexpr (std::is_same_v<typename fromUnit::Dim, typename toUnit::Dim>)
-      {
-         // Do conversion using Conversion struct
-         using conv = Conversion<fromUnit, toUnit>;
-         constexpr PrecisionType slope = GetSlope<conv>();
-         constexpr PrecisionType offset = GetOffset<conv>();
-         if constexpr (isDelta)
+         if constexpr (HasOffset<Conv> && !IsDelta) // Conversion provides offset
          {
-            if constexpr (inverse)
-            {
-               return (input / slope);
-            }
-            else
-            {
-               return (input * slope);
-            }
-
+            return input * Conv::slope + GetOffset<Conv>();
          }
-         else
+         else // Conversion does not provides offset
          {
-            return (input * slope) + offset;
+            return input * Conv::slope;
          }
       }
       else
       {
-         static_assert(sizeof(fromUnit) == -1, "No possible conversion for between these types.");
+         if constexpr (HasOffset<Conv> && !IsDelta) // Conversion provides offset
+         {
+            return (input - GetOffset<Conv>()) / Conv::slope;
+         }
+         else // Conversion does not provides offset
+         {
+            return input / Conv::slope;
+         }
       }
+   }
+
+
+   /// @brief Convert from one unit to another
+   /// @tparam From Unit to convert from
+   /// @tparam To Unit to convert to
+   /// @tparam IsDelta Bool indicating if the unit being converted is sole unit of the dimension 
+   ///   and is in the numerator (False), otherwise True
+   /// @tparam Inverse Whether to use the inverse of the conversion traits
+   /// @param input value to convert
+   /// @return converted value
+   template<typename From, typename To, bool IsDelta = false, bool Inverse = false>
+   requires (IsDelta || !Inverse)
+   constexpr PrecisionType Convert(PrecisionType input) 
+   {
+
+      // Checks on BaseDimension ensure a conversion to/from Primary exists.
+      // No need to further constrain this function.
+
+      if constexpr (std::is_same_v<From, To>) // Conversion to self
+      {
+         return input;
+      }
+      else if constexpr (HasConversion<From, To>) // Direct conversion exists
+      {
+         return ConvertImpl<Conversion<From, To>, IsDelta, Inverse>(input);
+      }
+      else if constexpr (HasConversion<To, From>) // Inverse direct conversion exists
+      {
+         return ConvertImpl<Conversion<To, From>, IsDelta, !Inverse>(input);
+      }
+      else // No direct conversion exists, fall back to primary
+      {
+         #ifdef REQUIRE_CONVERSIONS
+            static_assert(sizeof(fromUnit) == -1, "No specialized conversion found. See compiler output for more details");
+         #endif
+         // Temporary variable is NOT created to allow full constexpr behavior in situations where input itself is constexpr
+         return Convert<typename From::Primary, To, IsDelta, Inverse>
+         (
+            Convert<From, typename From::Primary, IsDelta, Inverse>(input)
+         );
+      }
+
    }
 
    /// @brief Convert the given value when units are cancelled out